backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
bilibili-api-python>=16.3.0
yt-dlp>=2024.12.0
loguru>=0.7.0
pydantic>=2.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
